apps/backend/researcher-service/pyproject.toml
[project]
name = "researcher-service"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "fastapi",
  "uvicorn",
  "sqlalchemy",
  "psycopg2-binary",
  "python-jose[cryptography]",
  "sqlglot",
  "redis",
  "pydantic"
]
